Moving and re-arranging the react app structure.
